+++ cloudant-client/src/main/java/com/cloudant/client/internal/util/CloudFoundryService.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2016 IBM Corp. All rights reserved.
+ * Copyright © 2016, 2018 IBM Corp. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
  * except in compliance with the License. You may obtain a copy of the License at
@@ -14,6 +14,8 @@
 
 package com.cloudant.client.internal.util;
 
+import com.google.gson.annotations.SerializedName;
+
 import java.net.URL;
 
 
@@ -24,6 +26,13 @@ public class CloudFoundryService {
 
     public static class CloudFoundryServiceCredentials {
 
-        public URL url;
+        @SerializedName("apikey")
+        public String apikey;
+        @SerializedName("host")
+        public String host;
+        @SerializedName("username")
+        public String username;
+        @SerializedName("password")
+        public String password;
     }
 }