+++ cloudant-client/src/test/java/com/cloudant/tests/HttpIamTest.java
@@ -22,8 +22,10 @@ import static com.cloudant.tests.util.MockWebServerResources.IAM_TOKEN;
 import static com.cloudant.tests.util.MockWebServerResources.IAM_TOKEN_2;
 import static com.cloudant.tests.util.MockWebServerResources.OK_IAM_COOKIE;
 import static com.cloudant.tests.util.MockWebServerResources.OK_IAM_COOKIE_2;
+import static com.cloudant.tests.util.MockWebServerResources.hello;
 import static com.cloudant.tests.util.MockWebServerResources.iamSession;
 import static com.cloudant.tests.util.MockWebServerResources.iamSessionUnquoted;
+import static com.cloudant.tests.util.MockWebServerResources.iamTokenEndpoint;
 import static org.hamcrest.CoreMatchers.anyOf;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -35,6 +37,7 @@ import com.cloudant.client.org.lightcouch.CouchDbException;
 import com.cloudant.http.Http;
 import com.cloudant.tests.extensions.MockWebServerExtension;
 import com.cloudant.tests.util.IamSystemPropertyMock;
+import com.cloudant.tests.util.MockWebServerResources;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
@@ -61,12 +64,8 @@ public class HttpIamTest {
     @RegisterExtension
     public MockWebServerExtension mockIamServerExt = new MockWebServerExtension();
 
-    final static String hello = "{\"hello\":\"world\"}\r\n";
-    final static String iamTokenEndpoint = "/identity/token";
-
     public MockWebServer mockWebServer;
     public MockWebServer mockIamServer;
-
 
     /**
      * Before running this test class setup the property mock.
@@ -114,33 +113,7 @@ public class HttpIamTest {
         String response = c.executeRequest(Http.GET(c.getBaseUri())).responseAsString();
         assertEquals(hello, response, "The expected response should be received");
 
-        // cloudant mock server
-        // assert that there were 2 calls
-        RecordedRequest[] recordedRequests = takeN(mockWebServer, 2);
-
-        assertEquals("/_iam_session",
-                recordedRequests[0].getPath(), "The request should have been for /_iam_session");
-        assertThat("The request body should contain the IAM token",
-                recordedRequests[0].getBody().readString(Charset.forName("UTF-8")),
-                containsString(IAM_TOKEN));
-        // the request should have the cookie header
-        assertEquals("/",
-                recordedRequests[1].getPath(), "The request should have been for /");
-        // The cookie may or may not have the session id quoted, so check both
-        assertThat("The Cookie header should contain the expected session value",
-                recordedRequests[1].getHeader("Cookie"),
-                anyOf(containsString(iamSession(EXPECTED_OK_COOKIE)),
-                        containsString(iamSessionUnquoted(EXPECTED_OK_COOKIE))));
-
-        // asserts on the IAM server
-        // assert that there was 1 call
-        RecordedRequest[] recordedIamRequests = takeN(mockIamServer, 1);
-        assertEquals(iamTokenEndpoint,
-                recordedIamRequests[0].getPath(), "The request should have been for " +
-                        "/identity/token");
-        assertThat("The request body should contain the IAM API key",
-                recordedIamRequests[0].getBody().readString(Charset.forName("UTF-8")),
-                containsString("apikey=" + IAM_API_KEY));
+        MockWebServerResources.assertMockIamRequests(mockWebServer, mockIamServer);
     }
 
     /**
@@ -170,33 +143,7 @@ public class HttpIamTest {
         String response = c.executeRequest(Http.GET(c.getBaseUri())).responseAsString();
         assertEquals(hello, response, "The expected response should be received");
 
-        // cloudant mock server
-        // assert that there were 2 calls
-        RecordedRequest[] recordedRequests = takeN(mockWebServer, 2);
-
-        assertEquals("/_iam_session",
-                recordedRequests[0].getPath(), "The request should have been for /_iam_session");
-        assertThat("The request body should contain the IAM token",
-                recordedRequests[0].getBody().readString(Charset.forName("UTF-8")),
-                containsString(IAM_TOKEN));
-        // the request should have the cookie header
-        assertEquals("/",
-                recordedRequests[1].getPath(), "The request should have been for /");
-        // The cookie may or may not have the session id quoted, so check both
-        assertThat("The Cookie header should contain the expected session value",
-                recordedRequests[1].getHeader("Cookie"),
-                anyOf(containsString(iamSession(EXPECTED_OK_COOKIE)),
-                        containsString(iamSessionUnquoted(EXPECTED_OK_COOKIE))));
-
-        // asserts on the IAM server
-        // assert that there was 1 call
-        RecordedRequest[] recordedIamRequests = takeN(mockIamServer, 1);
-        assertEquals(iamTokenEndpoint,
-                recordedIamRequests[0].getPath(), "The request should have been for " +
-                        "/identity/token");
-        assertThat("The request body should contain the IAM API key",
-                recordedIamRequests[0].getBody().readString(Charset.forName("UTF-8")),
-                containsString("apikey=" + IAM_API_KEY));
+        MockWebServerResources.assertMockIamRequests(mockWebServer, mockIamServer);
     }
 
     /**